Create a section as a part of test content

Create a section as a part of test content

Usage

section(
  content,
  n_variants = 1,
  seed_number = NULL,
  id = NULL,
  by = "variants",
  selection = 0,
  title = character(0),
  time_limits = NA_integer_,
  visible = TRUE,
  shuffle = FALSE,
  max_attempts = NA_integer_,
  allow_comment = TRUE
)

Arguments

content

string; vector of Rmd, md, xml files, task- or section-objects

n_variants

integer; number of variants to create from Rmd files

seed_number

integer vector, optional; seed numbers to reproduce the result of calculations

id

string, optional; identifier of the assessment section

by

"variants" or "files"; the type of the test structure; "variants" by default

selection

numeric, optional; defines how many children of the section are delivered in test

title

string, optional; title of the section

time_limits

integer, optional; controls the amount of time a candidate is allowed for this part of the test

visible

boolean, optional; if TRUE it shows this section in hierarchy of test structure; default TRUE

shuffle

boolean, optional; is responsible to randomize the order in which the assessment items and subsections are initially presented to the candidate; default FALSE

max_attempts

numeric, optional; enables the maximum number of attempts, that candidate is allowed to pass in this section

allow_comment

boolean, optional; enables to allow candidate to leave comments in each question of the section; TRUE by default

Value

object of AssessmentSection-class