Create test for LMS Opal

Create test for LMS Opal

Usage

test4opal(
  content,
  identifier = NULL,
  title = NULL,
  navigation_mode = "nonlinear",
  submission_mode = "individual",
  time_limits = NULL,
  max_attempts = NULL,
  allow_comment = TRUE,
  rebuild_variables = TRUE,
  files = NULL,
  show_test_time = FALSE,
  calculator = "scientific-calculator",
  mark_items = FALSE,
  keep_responses = FALSE
)

Arguments

content

list contains AssessmentSection objects

identifier

string; identifier if the test file

title

string, optional; file title

navigation_mode

string, optional; determines the general paths that the candidate may have during exam; two mode options are possible: linear "linear" - candidate is not allowed to return to the previous questions; nonlinear- candidate if free to navigate, used by default

submission_mode

string, optional; determines when the candidate's responses are submitted for response processing; two mode options are possible: individual - submit candidates' responses on an item-by-idem basis, used by default; simultaneous - candidates - responses are submitted all together by the end of the test

time_limits

numeric, optional; controls the amount of time a candidate is given for the test

max_attempts

numeric, optional; enables the maximum number of attempts, that candidate is allowed to pass

allow_comment

boolean, optional; enables to allow candidate to leave comments in each question, TRUE by default

rebuild_variables

boolean, optional; enables to recalculate variables and reshuffle the order of choices for each item-attempt

files

string vector, optional; paths to files, which will be accessible to candidate during the test/exam

show_test_time

boolean, optional; determines to show candidate elapsed processing time without time limit; default FALSE

calculator

string, optional; determines to show to candidate calculator; possible values: simple-calculator or scientific-calculator, the lase one is assigned by default

mark_items

boolean, optional; determines to allow candidate marking of questions, default TRUE

keep_responses

boolean, optional; determines to save candidate's answers of the previous attempt, default FALSE